// contracts/generated/typescript/models/EntityAlias.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Anby Wiki API
 * Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。 本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。 前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。 Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface EntityAlias
 */
export interface EntityAlias {
    /**
     *
     * @type {string}
     * @memberof EntityAlias
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EntityAlias
     */
    language: string;
    /**
     *
     * @type {string}
     * @memberof EntityAlias
     */
    alias: string;
    /**
     *
     * @type {EntityAliasAliasTypeEnum}
     * @memberof EntityAlias
     */
    aliasType: EntityAliasAliasTypeEnum;
}


/**
 * @export
 */
export const EntityAliasAliasTypeEnum = {
    Common: 'common',
    Historical: 'historical',
    Abbreviation: 'abbreviation',
    Import: 'import'
} as const;
export type EntityAliasAliasTypeEnum = typeof EntityAliasAliasTypeEnum[keyof typeof EntityAliasAliasTypeEnum];


/**
 * Check if a given object implements the EntityAlias interface.
 */
export function instanceOfEntityAlias(value: object): value is EntityAlias {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('language' in value) || value['language'] === undefined) return false;
    if (!('alias' in value) || value['alias'] === undefined) return false;
    if ((!('aliasType' in (value as Record<string, any>)) && !('alias_type' in (value as Record<string, any>))) || ((value as Record<string, any>)['aliasType'] === undefined && (value as Record<string, any>)['alias_type'] === undefined)) return false;
    return true;
}

export function EntityAliasFromJSON(json: any): EntityAlias {
    return EntityAliasFromJSONTyped(json, false);
}

export function EntityAliasFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntityAlias {
    if (json == null) {
        return json;
    }
    return {

        'id': json['id'],
        'language': json['language'],
        'alias': json['alias'],
        'aliasType': json['alias_type'],
    };
}

export function EntityAliasToJSON(json: any): EntityAlias {
    return EntityAliasToJSONTyped(json, false);
}

export function EntityAliasToJSONTyped(value?: EntityAlias | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'id': value['id'],
        'language': value['language'],
        'alias': value['alias'],
        'alias_type': value['aliasType'],
    };
}