// contracts/generated/typescript/models/EntityDetail.ts
export interface EntityDetail {
    /**
     *
     * @type {string}
     * @memberof EntityDetail
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EntityDetail
     */
    wikiId: string;
    /**
     *
     * @type {string}
     * @memberof EntityDetail
     */
    canonicalKey: string;
    /**
     *
     * @type {EntityDetailStatusEnum}
     * @memberof EntityDetail
     */
    status: EntityDetailStatusEnum;
    /**
     *
     * @type {string}
     * @memberof EntityDetail
     */
    mergedIntoEntityId?: string;
    /**
     *
     * @type {EntityTypeSummary}
     * @memberof EntityDetail
     */
    entityType: EntityTypeSummary;
    /**
     *
     * @type {Array<EntityLabel>}
     * @memberof EntityDetail
     */
    labels: Array<EntityLabel>;
    /**
     *
     * @type {Array<EntityAlias>}
     * @memberof EntityDetail
     */
    aliases: Array<EntityAlias>;
    /**
     *
     * @type {Date}
     * @memberof EntityDetail
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof EntityDetail
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const EntityDetailStatusEnum = {
    Active: 'active',
    Merged: 'merged',
    Deleted: 'deleted'
} as const;
export type EntityDetailStatusEnum = typeof EntityDetailStatusEnum[keyof typeof EntityDetailStatusEnum];